Retry database connection
When making a new connection to postgres, retry 10 times (bail out
after 20 seconds).  This mitigates against the case where the server
tries to connect to the database, but it is not yet available.

This may happen with `rake docker_up` where the server is set up
faster than the postgres container.