[#109] Implemented forwarding Kea command
Added new GRPC command to the Agent, to forward the Kea command to the
Kea instance over HTTP. The command name is ForwardToKeaOverHttp to
highlight that it is intended for sending the commands over Kea's
Control Agent (HTTP). In the future it is envisaged that we will also
have ForwardToKeaOverUnix which will be for forwarding the command over
the UNIX domain socket. This future command will lack the service list
and the response will be slightly different, therefore it makes sense
to split them into two commands.