[#165] Multiple subnets returned for prefix
In some cases we might have multiple subnets with the same prefix in the
database. Besides the fact we currently don't have a mechanism to merge
the subnets into one, in some deployments multiple subnets with the same
prefix may exist in different network segments. Kea doesn't support it
but we leave the door open.