[#170] Replace CtrlPort with AccessPoints
Replace CtrlAddress, CtrlPort, CtrlKey in the dbmodel.App with
AccessPoints.  Touches some code that interfaces with the dbmodel
package, and a lot of test code updates.

This requires a different way of comparing applications in
'database/model/service_test.go' because 'ElementsMatch' no longer
works on the new structure.

The various GetMachine and GetService are implemented suboptimal
in order to retrieve the access points on related apps.  This should
be fine for now, but should be improved when we deal with a lot of
apps.