[#165] Replaced app_to_subnet w/ local_subnet
The app_to_subnet table was renamed as it is meant to hold more info
that a simple M:N relationship between subnets and apps. Apart from the
rename the structures used in queries were also modified to make it easy
to access the information stored in local_subnet table.