[#64] Introduce agent.ForwardToNamedStats
Introduce a Stork agent function to forward a statistics request to
the named daemon statistics-channel.  The statistics-channel access
point is, just like with Kea Control Agent, represented by an URL.

This reuses the CAClient.  The CAClient has a Kea (Control Agent)
specific name but should be renamed because it is now also used
for the BIND 9 named statistics-channel.