[#214] Remove dangling hosts/subnets
If a host or subnet is no longer associated with any app as a result of
deletion of an app, the subnet/host is also removed. In the future we
may actually keep such dangling objects assuming they can be once again
associated with an app manually. Currently, there is no way to do it, so
the most reasonable approach seems to be to delete them.