[#64] Add query hit ratio
This is what we want when we talk about cache hit ratio. This is
the number of *queries* that are retrieved from the cache divided
by the total number of queries. The Cache Hit Ratio, is the number
of cache accesses that resulted in a cache hit, divided by the
total number of cache requests. This is a different statistic, and
since the number of cache misses never increases, seem to be broken.

Other changes:
- Change subsystem from 'cache' to 'view' (more appropriate)
- Remove Cache Hit Ratio from Stork and grafana dashboards