[#251] Correct HA state retrieving
Created new function which retrieves all HA states to which the given
daemon belongs. Still, only one HA state is possible but this function
is generic to support multiple HA states per daemons in the future.
This also corrects a bug, whereby we could fail to return HA state if there
was a non-HA service associated with the daemon.