[#446] Cancel puller during action function
Pullers are temporarily stopped for a period of time when puller action
is being run. This guarantees that the next action is triggered a certain
period of time after the current action has finished. This is important
when actions take a lot of time, e.g. adding new machine with many subnets.