[#477] Fixed app rename dialog box closing
It was indicated in the review comments that closing an app rename dialog
box with an escape caused an issue whereby one was unable to reopen the
dialog box without reloading the page. The issue was fixed by hooking up
on the onHide event of the dialog box and emitting an event to the parent
component in all cases when the dialog box is closed.

In addition, the keyUp event in the dialog box now handles pressing an
escape or enter during editing the name.