[#676] Fixed broken database migration 37
Invalid order of operations in the migration caused the Stork version 1.0
to fail to start when there were host reservations in the database. The
new columns have non-null contraints which are violated when local_host
entries exist. The local_host entries should be deleted along with the
host table entries but they aren't because of the missing CASCADE action.
The fix introduces the CASCADE action before the hosts deletion.