[#1784] Improved communication with stats
The gRPC API and the BIND9 HTTP client are better suited to handle long
lasting calls to the stats channel. The client avoids making calls to
the root path of the stats channel, and thus avoid fetching zones.
The gRPC call for ForwardToNamedStats now uses a set of parameters
instead of URL to specify the stats channel host, port and request type.
The request types are represented by an enum, so it is harder to make
a mistake calling invalid endpoint.