[#1956] Add app types filter on Apps list page
This change adds a MultiSelect drop-down component above the table on
the Apps page.  The drop-down allows users to select which kind of app
they want to see from the power set of {Kea, BIND9, PowerDNS}.  Changing
the drop-down causes an event which is debounced for 450ms and
deduplicated before causing a refresh of the apps table.  After
debouncing and deduplicating, the contents of the event are ignored,
because changing the rest of this page to be a reactive pipeline would
be a significant rewrite that would almost certainly collide with the
PrimeNG 17 -> 19 upgrade that Piotrek is working on.