[#2030] Use route-level code splitting for swagger
This moves the /swagger-ui route to a separate module.  It reduces the
main JS bundle size by 3 MB and improves First Contentful Paint from 10s
to 8s.  Moving it to a separate module allows Angular to avoid loading
the code if the user never navigates to that route, which is particulary
important for Swagger because it's loading a whole copy of React.