[#2055] Add first draft memfile snooper
In order to generate test data for the memfile snooper, I added
configurations to the demo which store DHCPv6 leases in a memfile as
part of the HA setup.

The memfile snooper is a generic struct with attached methods to parse
and emit either Lease4 or Lease6 structs.  It polls the lease memfile in
a 250ms loop (for now) so that no additional dependency for
kqueue/epoll/etc. is necessary.

I have included a number of tests which prove it works correctly in
several situations, although I'm sure I've missed some details.